fix three more warnings
This call site is reachable on: 'iOS' 12.2 and later, 'maccatalyst' 12.2 and later.
'UIGraphics.BeginImageContextWithOptions(CGSize, bool, NFloat)' is unsupported on: 'ios' 17.0 and later, 'maccatalyst' 17.0 and later.

This call site is reachable on: 'iOS' 12.2 and later, 'maccatalyst' 12.2 and later.
'UIGraphics.GetImageFromCurrentImageContext()' is unsupported on: 'ios' 17.0 and later, 'maccatalyst' 17.0 and later.

This call site is reachable on: 'iOS' 12.2 and later, 'maccatalyst' 12.2 and later.
'UIGraphics.EndImageContext()' is unsupported on: 'ios' 17.0 and later, 'maccatalyst' 17.0 and later.